feat(setup): backend review

+++ src/fs/roots.rs
@@ -7,9 +7,9 @@ use crate::state::AppState;
 #[derive(Serialize)]
 pub(super) struct Root {
     name: String,
-    total: u64,     // capacity of the filesystem holding this root
-    used: u64,      // total minus available
-    available: u64, // free space usable without privileges
+    total: u64,
+    used: u64,
+    available: u64,
 }
 
 pub(super) async fn list_roots(State(state): State<AppState>) -> Json<Vec<Root>> {
@@ -19,9 +19,9 @@ pub(super) async fn list_roots(State(state): State<AppState>) -> Json<Vec<Root>>
         .map(|(name, path)| {
             let (total, used, available) = match statvfs(path.as_path()) {
                 Ok(s) => {
-                    let unit = s.fragment_size() as u64;
-                    let total = s.blocks() as u64 * unit;
-                    let available = s.blocks_available() as u64 * unit;
+                    let unit = s.fragment_size();
+                    let total = u64::from(s.blocks()) * unit;
+                    let available = u64::from(s.blocks_available()) * unit;
                     (total, total.saturating_sub(available), available)
                 }
                 Err(_) => (0, 0, 0),
